+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.0]
+
+### Added
+
+* Add "select all targets" command with a button in the main tree view.
+
+### Changed
+
+* Replace `.png` icons with `vscode.ThemeIcon` icons.
+
+### Fixed
+
+* Fix app sideload task requirements installation on macOS.
+
 ## [0.3.4]
 
 ### Fixed